+++ CHANGELOG.md
@@ -3,8 +3,7 @@
 ## New Features
 
 - PR #1194  Implement overloads for CUDA atomic operations
-
-...
+- PR #1292 Implemented Bitwise binary ops AND, OR, XOR (&, |, ^)
 
 ## Improvements